Link UI To Core
So, I've finally linking up the UI to the core - this build will most-likely be full of ***tons*** of bugs, but that's ok since the UI is getting replaced with WPF soon anyway.

The Quality Code Control still hasn't taken very much effect and I still haven't completely perfermormed that.

Also made a few changes to the CONTRIBUTING.md file.